refactor: extract email endpoints to their own file
